remove one layer of coconut in nym-cli
